+++ catalyst/contrib/scripts/collect_env.py
@@ -1,5 +1,5 @@
 # flake8: noqa
-# @TODO: recheck codestyle later
+# @TODO: code formatting issue for 20.07 release
 """
 This script outputs relevant system environment info.
 Diagnose your system and show basic information.