+++ catalyst/dl/__main__.py
@@ -8,7 +8,11 @@ COMMANDS = OrderedDict([("init", init), ("run", run), ("trace", trace)])
 
 
 def build_parser() -> ArgumentParser:
-    """@TODO: Docs. Contribution is welcome."""
+    """Builds parser.
+
+    Returns:
+        parser
+    """
     parser = ArgumentParser(
         "catalyst-dl", formatter_class=RawTextHelpFormatter
     )