+++ catalyst/utils/checkpoint.py
@@ -1,3 +1,5 @@
+# flake8: noqa
+# @TODO: code formatting issue for 20.07 release
 from typing import Dict, Union
 from collections import OrderedDict
 import os