+++ catalyst/utils/tests/registery_foo.py
@@ -1,3 +1,7 @@
+# flake8: noqa
+# @TODO: code formatting issue for 20.07 release
+
+
 def foo(a, b):
     """@TODO: Docs. Contribution is welcome."""
     return {"a": a, "b": b}