+++ tests/_tests_cv_classification_experiment_registry/test1/experiment.py
@@ -1,3 +1,4 @@
+# flake8: noqa
 from collections import OrderedDict
 
 from catalyst.contrib.data.transforms import Compose, Normalize, ToTensor